refactor: change to re-ducks composition
