--- src/actions.ts
@@ -1,3 +1,0 @@
-import { Actions as CounterActions } from './modules/counter/module'; // TODO: export interface/type
-
-export type Actions = CounterActions;
