+++ src/stores/actions.ts
@@ -1,0 +1,2 @@
+import { Actions as CounterActions } from '../modules/counter/actions';
+export type Actions = CounterActions;
