+++ src/stores/reducers.ts
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import { reducer as counterReducer } from './modules/counter';
+import { reducer as counterReducer } from '../modules/counter';
 
 export const rootReducer = combineReducers({
   counter: counterReducer,
