+++ src/stores/store.ts
@@ -1,7 +1,7 @@
 import { createStore } from 'redux';
 import { rootReducer } from './reducers';
-import { nodeEnv } from './env';
-import { State as CounterState } from './modules/counter/module'; // TODO: export interface/type
+import { nodeEnv } from '../env';
+import { State as CounterState } from '../modules/counter/reducers';
 
 export default createStore(
   rootReducer,
